Added intersection_num program
this class contains a method that takes as input a word (in our gernerating set) and returns the self-intersection number of the corresponding closed curve.

This program calculates the self-intersection number by calling the auxilliary pgrogram "self-ints", which is a modified version of Alden Walker's program "branched". Thus in order to work, this program must be run on a machine where the program "self-ints" is installed in the $PATH.